implement (untested) negative delay in hz display
Writing a very long explanation here as I think these changes are
somewhat brittle.

The delay value is stored in hzSpawnDelay, which is thankfully detached
from basically all other hz computations. Thus, it's actually simplest
to keep all other functionality as it is, i.e. the taps/speed metrics
only act on inputs made while a piece is active.

Negative delay is currently limited to -3. This choice is somewhat
arbitrary and could reasonably be extended in the future, but I think
this is pretty realistic for most people who do care about timing. To
support other entry delay frames, add a call to checkNegativeDelay in
the relevant branch, making sure to set X to the correct value. Negative
values themselves are stored as the magnitude ORed with $80, so e.g. -2
is represented as $82.

One tricky part is that the entry delay pipeline techincally has
multiple paths. The normal path goes through playState_receiveGarbage
once and playState_spawnNextTetrimino twice as the last three frames,
but the game can spend many more frames in playState_spawnNextTetrimino
if the spawnDelay parameter is set, which is done in the tspin and debug
modes. There is extra logic to handle positive spawnDelay, but any other
changes to how entry delay works might require further refactoring.