replace switch_s_plus_2a with branch macro
Uses a macro with destinations as args to create a high bytes and a low bytes
table for each destination offset by -1, then uses the stack and rts for the
actual branching.

Macro expands to 11-12 bytes per branch with nothing centralized, compared to
the previous 5-6 bytes per branch to a centralized routine.

Reduces cycle count from 54-57 cycles per branch to 23-26, with the variance
from possible page boundary crossing and if the destination is in zero page.
Normal game logic involves 16 branches per frame.

Supports up to 16 destinations but can be easily extended, the most in use is
the playstate branch with 12 destinations.

Previous:
        lda dest                ; 3/4
        jsr switch_s_plus_2a    ; 6

switch_s_plus_2a:
        asl a                   ; 2
        tay                     ; 2
        iny                     ; 2
        pla                     ; 4
        sta switchTmp1          ; 3
        pla                     ; 4
        sta switchTmp2          ; 3
        lda (switchTmp1),y      ; 5/6
        tax                     ; 2
        iny                     ; 2
        lda (switchTmp1),y      ; 5/6
        sta switchTmp2          ; 3
        stx switchTmp1          ; 3
        jmp (switchTmp1)        ; 5
                                ; 54-57

New:
        ldx dest                ; 3/4
        lda hiBytes,x           ; 4/5
        pha                     ; 3
        lda loBytes,x           ; 4/5
        pha                     ; 3
        rts                     ; 6
                                ; 23-26