remove unneeded legacy tables
orientationToSpriteTable is used to map the spawnId to the arbitrary sprite ID
for the next piece.  Reordering the sprites to match the game's piece ID order
and reusing tetriminoTypeFromOrientation, this table can be removed.

spawnOrientationFromOrientation may have been created in development when the
RNG produced the orientation ID instead of the piece ID, but random spawn
orientations weren't desired.  As it's currently being used to convert the spawn
ID to the same spawn ID, it can be removed.

Reduces PRG usage by 38 bytes