branch differently for fewer destinations
Throws warning when single destination is defined but optimizes to jmp
instruction anyway.

For 2 to 4 destinations, uses a decrementing x register, branch
instructions and jmp instructions to save a minimum of 7 cycles, up to 18.

Rom space is saved for 1-3 destinations.  In the case of 4 destinations,
it uses an additional 2 bytes compared to the rts method.

How each option expands:

1 destination: 3 cycles
    jmp a1

2 destinations: maximum 11 cycles to a0
    ldx dest
    beq addr0
    jmp a1
addr0:
    jmp a0

3 destinations: maximum 15 cycles to a1
    ldx dest
    beq addr0
    dex
    beq addr1
    jmp a2
addr1:
    jmp a1
addr0:
    jmp a0

4 destinations: maximum 19 cycles to a2
    ldx dest
    beq addr0
    dex
    beq addr1
    dex
    beq addr2
    jmp a3
addr2:
    jmp a2
addr1:
    jmp a1
addr0:
    jmp a0

5+ destinations: 23-26 cycles
    ldx dest
    lda hyBytes,x
    pha
    lda loBytes,x
    pha
    rts
hiBytes:
    .byte >a0,>a1,>a2,>a3,>a4
hiBytes:
    .byte <a0,<a1,<a2,<a3,<a4